% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acaps_npi_data.R
\name{download_acaps_npi_data}
\alias{download_acaps_npi_data}
\title{Download ACAPS non-pharmaceutical interventions data}
\usage{
download_acaps_npi_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data, organized by intervention. It
    includes a \code{timestamp} variable indicating the time of data
    retrieval.
}
\description{
Downloads non-pharmaceutical interventions (NPI) data related to Covid-19
from the ACAPS governmental measures database
(\url{https://www.acaps.org/covid19-government-measures-dataset}).
}
\examples{
df <- download_acaps_npi_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(country) \%>\%
  dplyr::summarise(number_of_interventions = dplyr::n()) \%>\%
  dplyr::arrange(-number_of_interventions)

}
