% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_apple_mtr_data.R
\name{download_apple_mtr_data}
\alias{download_apple_mtr_data}
\title{Download Apple Mobility Trend Reports Data}
\usage{
download_apple_mtr_data(
  type = "country",
  url = NULL,
  silent = FALSE,
  cached = FALSE
)
}
\arguments{
\item{type}{The type of data that you want to retrieve. Can be any subset of
\itemize{
 \item{"country": }{Mobility trends by country.}
 \item{"country_region": }{Mobility trends by country regions as classified by Apple (only available for some countries).}
 \item{"country_city": }{Mobility trends by city as classified by Apple (only available for some countries).}
}}

\item{url}{The URL of the Apple Mobility Trend Reports is burried in
Java Script, making it non-trivial to parse. The Github repository
of this package contains an R script that uses the 'RSelenium'
package and headless browsing to scrape the URL. The scraped URL
is included in the daily updated data contained in this repository.
By default, the function uses this cached URL to retrieve the data.
You can, however, pass your own URL by setting this variable.
Defaults to \code{NULL} to trigger loading the cached URL.}

\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data. It includes a \code{timestamp}
    variable indicating the time of data retrieval.
}
\description{
Downloads Mobility Trends Reports provided by Apple related to Covid-19
(\url{https://www.apple.com/covid19/mobility}).
}
\examples{
df <- download_apple_mtr_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarize(
    walking = mean(walking, na.rm = TRUE)
  ) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = walking)) +
  ggplot2::geom_line()

df <- download_apple_mtr_data(type = "country_city", silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::filter(iso3c == "DEU") \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = driving, color = city)) +
  ggplot2::geom_line()

}
