% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ecdc_covid19_data.R
\name{download_ecdc_covid19_data}
\alias{download_ecdc_covid19_data}
\title{Download ECDC Data on Covid-19 cases}
\usage{
download_ecdc_covid19_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data.
}
\description{
Downloads case data provided by the European Centre for Disease Prevention
and Control
(\url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}).
The data is updated daily and contains the latest available public data
on the number of new Covid-19 cases reported per day and per country.
}
\examples{
df <- download_ecdc_covid19_data(silent = TRUE, cached = TRUE)

df \%>\%
  dplyr::filter(iso3c == "ITA", !is.na(cases)) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = cases)) +
  ggplot2::geom_bar(stat = "identity", fill = "lightblue")

}
