% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_google_cmr_pdfs.R
\name{download_google_cmr_pdfs}
\alias{download_google_cmr_pdfs}
\title{Download Google Community Mobility Reports as PDF files}
\usage{
download_google_cmr_pdfs(pdf_dir, silent = FALSE)
}
\arguments{
\item{pdf_dir}{Path to the directory where the PDF files will be stored.}

\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}
}
\value{
A vector containing the filenames that were downloaded.
}
\description{
Downloads Google Community Mobility Reports
(\url{https://www.google.com/covid19/mobility/}).
As stated on this webpage, the "reports chart movement trends over time
by geography, across different categories of places such as retail and
recreation, groceries and pharmacies, parks, transit stations,
workplaces, and residential". Google prepares these reports to help
interestes parties to assess responses to social distancing guidance
related to Covid-19.
}
\examples{
\dontrun{
 download_google_cmr_pdfs(dir "~/gogle_cmr_pdfs", silent = TRUE)
}

}
