% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_owid_data.R
\name{download_owid_data}
\alias{download_owid_data}
\title{Download Our World in Data data}
\usage{
download_owid_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data.
}
\description{
Downloads data on testing, hospital occupation and vaccinations
as collected by the Our World in Data team from official reports.
You can find the source information for every country at
(\url{https://github.com/owid/covid-19-data/tree/master/public/data}).
The data accumulates tests over time. The definition of what constitutes
a test and the frequency of data collection vary across countries. The
vaccination data is currently only available based on ad hoc disclosures
by a small set of countries.
}
\examples{
df <- download_owid_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(iso3c) \%>\%
  tidyr::fill(total_tests) \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarise(tests = sum(total_tests, na.rm = TRUE)) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = tests)) +
  ggplot2::geom_line()

}
