% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_owid_testing_data.R
\name{download_owid_testing_data}
\alias{download_owid_testing_data}
\title{Download Our World in Data testing data}
\usage{
download_owid_testing_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data.
}
\description{
Downloads testing data as collected by the Our World in Data team from
official reports. You can find the source information for every country at
(\url{https://ourworldindata.org/covid-testing}).
The data accumulates tests over time. The definition of what constitutes
a test and the frequency of data collection vary across countries.
}
\examples{
df <- download_owid_testing_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(iso3c) \%>\%
  tidyr::fill(total_tests) \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarise(tests = sum(total_tests, na.rm = TRUE)) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = tests)) +
  ggplot2::geom_line()

}
