% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid19_spread.R
\name{plot_covid19_spread}
\alias{plot_covid19_spread}
\title{Visualize the spread of the Covid-19 pandemic}
\usage{
plot_covid19_spread(data = download_merged_data(cached = TRUE, silent =
  TRUE), type = "deaths", min_cases = ifelse(type == "deaths", 10,
  100), min_by_ctry_obs = 7, edate_cutoff = 30,
  data_date_str = format(lubridate::as_date(data$timestamp[1]),
  "\%B \%d, \%Y"), per_capita = FALSE, highlight = NULL,
  intervention = NULL)
}
\arguments{
\item{data}{The data frame to base the plot on. Should be a merged data
frame obtained by \link{download_merged_data} and defaults to
\code{download_merged_data(cached = TRUE, silent = TRUE)}.}

\item{type}{The statistic that you want to plot. Needs to be either "confirmed",
"deaths", or "revovered".}

\item{min_cases}{Defines the zero point of your X axis (the 'event date').
Defaults to 10 cases for deaths and 100 cases otherwise.}

\item{min_by_ctry_obs}{Limits the plot to countries that have at least that
many days of dater since and including the event date. Defaults to 7.}

\item{edate_cutoff}{The upper limit of the X axis in event days.
Defaults to 30.}

\item{data_date_str}{A date string to include in the annotation of the plot
giving the time when the data was pulled. Defaults to the time stemp of the
data. Note that you might run into issues with the default when running
this in a non-english locale. Consider setting it by hand then.}

\item{per_capita}{If \code{TRUE} data is being plotted as per capita measure
based on World Bank data. Defaults to \code{FALSE}.}

\item{highlight}{A character vector of ISO3c (ISO 3166-1 alpha-3) codes that
identify countries that you want to highlight. Using the
\code{gghighlight} package, these observations are highlighted by color
and labeled while the other are grayed out. In \code{NULL} (the default),
all countries are labeled. This can cause very cluttered plots.}

\item{intervention}{If not default \code{NULL} then this identifies the
intervention type that you want to be highlighted by a point marker.
Valid intervention types are based on the ACAPS government measure data
and include general lockdowns ('lockdown'), social distancing
('soc_dist'), movement restrictions ('mov_rest'), public health measures
('pub_health'), and social and economic measures ('soc_econ').}
}
\value{
A \link{ggplot2} object.
}
\description{
Provides a flexible visualization of the country-level Covid-19 spread,
inpired by the displays created by John Burn-Murdoch from the Financial
Times. Uses data from the Johns Hopkins University CSSE team
(\url{https://github.com/CSSEGISandData/COVID-19}) and the ACAPS governmental
measures database
(\url{https://www.acaps.org/covid19-government-measures-dataset}).
If your are overwhelmed with the options of the plot, explore them with
\code{shiny_covid19_spread()} and use the 'copy code for plot to clipboard'
option.
}
\examples{
plot_covid19_spread()

merged <- download_merged_data(cached = TRUE, silent = TRUE)
plot_covid19_spread(merged, highlight = "DEU", intervention = "lockdown")

plot_covid19_spread(merged, type = "recovered", min_by_ctry_obs = 10,
  edate_cutoff = 40,
  highlight = c("ITA", "ESP", "FRA", "DEU", "USA"), intervention = "soc_dist"
)

}
