% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_google_cmr_data.R
\name{scrape_google_cmr_data}
\alias{scrape_google_cmr_data}
\title{Scrape Google Community Mobility Report Data}
\usage{
scrape_google_cmr_data(
  daily_data = TRUE,
  pdf_dir = NULL,
  silent = FALSE,
  cached = FALSE
)
}
\arguments{
\item{daily_data}{Whether you want the code to extract daily data by
scanning the line graph figures that are included in the PDF. The
resulting data are noisy but should be accurate by +/- one day and
+/- 2 percent. Defaults to \code{TRUE}.}

\item{pdf_dir}{Path to the directory where the PDF files are stored.
If \code{NULL} (the default), the code will download the PDFs from
Google's website (see above). See \code{google_download_cmr_pdfs()}
for downloading the PDFs.}

\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
If \code{daily_data == FALSE}, a data frame containing the
    country level data. Otherwise, a list containing the country level data
    and the country-day level data.
}
\description{
Scrapes (and downloads) Google Covid-19 Community Mobility Reports
(\url{https://www.google.com/covid19/mobility/}).
As stated on this webpage, the "reports chart movement trends over time
by geography, across different categories of places such as retail and
recreation, groceries and pharmacies, parks, transit stations,
workplaces, and residential". Google prepares these reports ino PDF format
to help interestes parties to assess responses to social distancing guidance
related to Covid-19.
}
\details{
The data is currently parsed from PDF files that Google provides.
   Thus, the code will most likely break as soon Google is changing the
   PDF format of their files.
}
\examples{

df <- scrape_google_cmr_data(daily_data = FALSE, silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::select(iso3c, retail_recreation) \%>\%
  dplyr::arrange(retail_recreation)

lst <- scrape_google_cmr_data(silent = TRUE, cached = TRUE)
lst[[2]] \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarize(
    retail_recreation = mean(retail_recreation, na.rm = TRUE)
  ) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = retail_recreation)) +
  ggplot2::geom_line()


}
