% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_covid19_spread.R
\name{shiny_covid19_spread}
\alias{shiny_covid19_spread}
\title{Explore the spread of Covid-19 interactively}
\usage{
shiny_covid19_spread(plot_options = list(), ...)
}
\arguments{
\item{plot_options}{A named list containing parameters to
initialize the plot in the shiny app. All parameters of
\code{plot_covid19_spread()} are supported.}

\item{...}{Parameters that will be parsed on to \code{\link[shiny]{runApp}}.}
}
\description{
A shiny based web app that allows users to customize the
   \code{\link{plot_covid19_spread}} display. The display has been
   inpired by the displays created by John Burn-Murdoch for the Financial
   Times. You can use it to customize your \code{plot_covid19_spread()}
   display as it allows copying the plot generating code to the clipboard,
   thanks to the fine {rclipboard} package.
}
\details{
Uses data from the Johns Hopkins University CSSE team on the spread of the
SARS-CoV-2 virus and the Covid-19 pandemic
(\url{https://github.com/CSSEGISandData/COVID-19}), from the ACAPS
governmental measures database
(\url{https://www.acaps.org/covid19-government-measures-dataset}) and from
the World Bank (\url{https://data.worldbank.org}). See the
documentation of the separate download functions of the package for more
detail.
}
\examples{
\dontrun{
  shiny_covid19_spread()

  df <- download_merged_data(cached = TRUE)
  shiny_covid19_spread(
    plot_options = list(data = df, highlight = c("USA", "ESP", "ITA"))
  )
}
}
