% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidycovid19_data_sources}
\alias{tidycovid19_data_sources}
\title{Documentation of \code{tidycovid19} Data Sources}
\format{
An object of class \code{"data.frame"}.
}
\source{
Information has been collected by data providers (URLs included in
  data frame). Descriptions of data cleaning steps provided by the author
  of this package.
}
\usage{
data(tidycovid19_data_sources)
}
\description{
Brief info on each data source included in \code{tidycovid19}
}
\examples{
data(tidycovid19_data_sources)
print(tidycovid19_data_sources)

}
\keyword{datasets}
