% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidycovid19_variable_definitions}
\alias{tidycovid19_variable_definitions}
\title{Variable Definitions of \code{tidycovid19} Merged Data Frame}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(tidycovid19_variable_definitions)
}
\description{
List variable definitions and data sources for the data frame
created by \code{download_merged_data()}.
Merge with \code{data(tidycovid19_data_sources)} to get full information
for each variable, including description of data source and link to URL.
}
\examples{
vd <- tidycovid19_variable_definitions
ds <- tidycovid19_data_sources
data_info <- dplyr::left_join(vd, ds, by = c(var_source = "id"))
print(data_info)

}
\keyword{datasets}
