% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_google_cmr_data.R
\name{download_google_cmr_data}
\alias{download_google_cmr_data}
\title{Download Google Community Mobility Reports Data}
\usage{
download_google_cmr_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
The data, tidied at the country level
}
\description{
Downloads Google Community Mobility Report data
(\url{https://www.google.com/covid19/mobility/}).
As stated on this webpage, the "reports chart movement trends over time
by geography, across different categories of places such as retail and
recreation, groceries and pharmacies, parks, transit stations,
workplaces, and residential". Google prepares these reports to help
interested parties to assess responses to social distancing guidance
related to Covid-19.
}
\examples{
df <- download_google_cmr_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarize(
    retail_recreation = mean(retail_recreation, na.rm = TRUE)
  ) \%>\%
  ggplot2::ggplot(ggplot2::aes(x = date, y = retail_recreation)) +
  ggplot2::geom_line()

}
