% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_google_cmr_pdfs.R
\name{download_google_cmr_pdfs}
\alias{download_google_cmr_pdfs}
\title{Download Google Community Mobility Reports as PDF files}
\usage{
download_google_cmr_pdfs(
  pdf_dir,
  date_indexed = FALSE,
  all_dates = FALSE,
  silent = FALSE
)
}
\arguments{
\item{pdf_dir}{Path to the directory where the PDF files will be stored.}

\item{date_indexed}{When \code{TRUE}, the function first reads the date
contained in the PDF file names. If the data is newer than the last
date contained in the directory names in \code{pdf_dir}, it downloads
the new PDFs into a new directory named by date. If not, it downloads
nothing. If \code{FALSE} (the default), it simply downloads the current
PDFs to \code{pdf_dir} regardless of their date.}

\item{all_dates}{Currently, Google still seems to keep the PDFs from
previous dates around. If you set this to \code{TRUE}, the function will
try to download old PDFs as well and stores them in directories named by
date. Required \code{date_indexed == TRUE} and defaults to \code{FALSE}.}

\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}
}
\value{
The datestamp of the available PDFs.
}
\description{
Downloads Google Community Mobility Reports
(\url{https://www.google.com/covid19/mobility/}).
As stated on this webpage, the "reports chart movement trends over time
by geography, across different categories of places such as retail and
recreation, groceries and pharmacies, parks, transit stations,
workplaces, and residential". Google prepares these reports to help
interested parties to assess responses to social distancing guidance
related to Covid-19.
}
\examples{
\dontrun{
 download_google_cmr_pdfs(dir "~/gogle_cmr_pdfs", silent = TRUE)
}

}
