% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_oxford_npi_data.R
\name{download_oxford_npi_data}
\alias{download_oxford_npi_data}
\title{Download Oxford non-pharmaceutical interventions data}
\usage{
download_oxford_npi_data(silent = FALSE, cached = FALSE)
}
\arguments{
\item{silent}{Whether you want the function to send some status messages to
the console. Might be informative as downloading will take some time
and thus defaults to \code{TRUE}.}

\item{cached}{Whether you want to download the cached version of the data
from the {tidycovid19} Github repository instead of retrieving the
data from the authorative source. Downloading the cached version is
faster and the cache is updated daily. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the data, organized by intervention. It
    includes a \code{timestamp} variable indicating the time of data
    retrieval.
}
\description{
Downloads non-pharmaceutical interventions (NPI) data related to Covid-19
from the Oxford Covid-19 Government Response Tracker.
(\url{https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker}).
It currently only uses the policy measures from that data and tidies them
into long format, defining observations to be interventions and discarding
observations with \code{NA} and unchanged \code{0} measures.
}
\details{
The Oxford data is currently not included in the data frame
    produced by \code{download_merged_data()} as the ACAPS NPI data seem
    to be of better quality overall. See
    \href{https://joachim-gassen.github.io/2020/04/exploring-and-benchmarking-oxford-government-response-data/}{this blog post}
    for a discussion.
}
\examples{
df <- download_oxford_npi_data(silent = TRUE, cached = TRUE)
df \%>\%
  dplyr::group_by(country) \%>\%
  dplyr::summarise(number_of_interventions = dplyr::n()) \%>\%
  dplyr::arrange(-number_of_interventions)

}
